+++ src/main/java/domain/ladder/stick/RandomStickGenerator.java
@@ -1,4 +1,7 @@
-package domain.ladder;
+package domain.ladder.stick;
+
+import domain.ladder.stick.Stick;
+import domain.ladder.stick.StickGenerator;
 
 import java.util.Random;
