+++ src/main/java/domain/ladder/stick/Stick.java
@@ -1,4 +1,4 @@
-package domain.ladder;
+package domain.ladder.stick;
 
 public enum Stick {
