+++ src/main/java/domain/ladder/stick/StickGenerator.java
@@ -1,4 +1,4 @@
-package domain.ladder;
+package domain.ladder.stick;
 
 public interface StickGenerator {
