+++ src/test/java/domain/LadderTest.java
@@ -2,8 +2,8 @@ package domain;
 
 import domain.ladder.Height;
 import domain.ladder.Ladder;
-import domain.ladder.Stick;
-import domain.ladder.StickGenerator;
+import domain.ladder.stick.Stick;
+import domain.ladder.stick.StickGenerator;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
