+++ src/test/java/domain/LineTest.java
@@ -1,8 +1,8 @@
 package domain;
 
 import domain.ladder.Line;
-import domain.ladder.Stick;
-import domain.ladder.StickGenerator;
+import domain.ladder.stick.Stick;
+import domain.ladder.stick.StickGenerator;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
